Return HATEOAS style paths instead of tokens
HATEOAS can make APIs easier to use. Instead of asking users to pull out
a token and construct a URL, we can construct the URL for them. This
means they just need to append the URL to the base domain and need not
know anything else.

Preferably the base domain would also be included. This can be done with
a configuration object containing the host name.