Correctly omit next when limit is defaulted
When limit is defaulted, our (len >) check to know if there's a next
page doesn't work so we just include it. To correctly handle this case,
we need to know the real limit even after any defaulting.

I elected to let the library handle the defaulting now. The alternative
would mean another argument passed in to the withPage functions. I think
it's OK for the library to own this, and if clients care, they should be
passing &limit generally anyway.