Read limit, don't parse it as JSON
Fixes #3. As discussed there, the limit parameter is a simple number.
Parsing it as JSON works, but is superfluous. Reading it (via Read or
PathPiece) would function the same and be simpler, Read being simplest.

I chose to make a newtype for additional safety. Originally, I tried to
wrap a Natural, but in doing so I was alerted to a place where we call
(!!) with (limit - 1). The list we're indexing is guarded to be larger
then limit, making (!!) safe, but only for limit values greater than 0.

Therefore, I additionally assert that we're not given 0. And once I was
doing that kind of validation myself, there wasn't much Natural would be
giving me. And the operations we do with the underlying representations
are easier if we don't have to fromIntegral all over, so I kept it Int.