Correctly handle conditional previous-page link
The handling of this was incorrect with regard to the extra item
indicating presence of a Next page or Previous page.

To work correctly, we need to switch on the current position a few times
to ensure that:

- The "extra" item existing is only used to indicate a Next page if we
  were "traveling in that direction". Otherwise, it should be used to
  indicate a Previous page existing

- Being on a First or Next page can know a-priori if there's a Previous
  page. Similarly, being on a Last or Previous page can know if there's
  a Next page -- checking for the "extra item" in either of these cases
  will give the wrong answer half the time.

Since the current test suite only ever checked for conditionality of a
Previous link from a Last page, it happened to work. The failing case
was discovered when trying to use this library in an app, via it's test
suite, and backfilled here.