Add some dependency bounds
This is a bit nuanced, and I'm not sure if it's the best approach, so
here is some explanation:

Managing dependency bounds manually is an error-prone pain. Stack
supports a --pvp-bounds option to sdist (and upload, which uses it) to
automate the process. It can add lower bounds, upper bounds, or both. It
operates by looking at the version of every dependency in the resolver
in use and sets a >= lower bound on that version and a < upper bound on
the next major version.

See https://www.fpcomplete.com/blog/2015/09/stack-pvp/

We've decided to omit bounds in our committed source generally, and use
the --pvp-bounds both option when releasing to Hackage. This is a
conservative choice, adding tight bounds based on the latest LTS
resolver, which is what we typically prefer in stack.yaml.

Where this breaks down is when building the project, with those
automated pvp-bounds, against nightly. Since the "next major" upper
bound is relative to the current LTS resolver, it's likely that the
nightly resolver will have a newer version and not build. Worse, you
can't know this until you actually try to build from the package built
with sdist -- since bounds aren't committed in package.yaml.

The solution is simple but annoying: add back manual bounds that are
more relaxed, which is what this commit does.

Will this be just as error prone and annoying as not using the
pvp-bounds option at all? Would using pvp-bounds upper or lower only
work out better for us? I'm not sure, we'll see.