Support http-link-header-1.2
This update replaced `Link` with `IsURI uri => Link uri`, allowing
different URI-like types to be used. This means we cannot use both 1.1
(in current LTS) and 1.2 (just landed in nightly).

This PR creates a small module that exposes a URI-specific interface
used by the library. CPP can break tooling (brittany), so to make it
more tolerable, I try to isolate it to a small Compat module like this,
vs peppering it around in the "actual code".

As an alternative, we could have raise the lower bound on
http-link-header to >= 1.2.

That version isn't in LTS yet, so we would ourselves need to release as
a major version bump or we would be pulled out as incompatible.

So we would then have to either use nightly as our default resolver
here, or bring newer http-link-header in as an extra-dep. Folks using us
(including ourselves) would either have to remain behind for now, use
nightly themselves, or bring in http-link-header and us in as
extra-deps.

Remaining behind is probably fine for our users (that's sort of the
point of choosing LTS), but for us as maintainers having to port a
potential bugfix back to the 1.1x branch is not the kind of thing we
normally set ourselves up to need to do.

And all of the resolver/extra-dep shuffling would ideally be rewound
once all the things moved back into LTS again. Because these activities
spider out into our users, it seems worth avoiding.

CPP is a slippery-slope of maintenance burden that I normally do not
like to take on, but the trade-offs in the alternative are too much for
me in this case.