Re-order mkYesod/instance-Yesod
Yesod has a superclass constraint on RenderRoute, which is provided by
mkYesod TemplateHaskell. It appears that GHC 9 (just hit nightly)
requires the TH that generates that class appear before the instance
that needs it.

I honestly would've always expected that to be the case, but this code
was fine before. Oh well.